Removed unnecessary prints. Fixed a bug(?)
Removed prints on MMap and Global, on trace_load.
Removed print Global, on trace_store and trace_modify.
Do not need this information on mapvisual.

Removed information printed on soname and function name.
Another information that is not need on mapvisual.
Removing function name, makes a big step to a fast mapvisual.
Not only reduces filesize extremely.

(for tangledaccess N=2048 (personal benchmark)
under --read-debug=yes and --read-var-info=yes flags
gleipnir.out dropped from 4.5Mb to 3.4Mb or ~25% reduced file.)

(for part of frames in slambench, in raycast() called from
raycastKernel()
under --read-debug=yes and --read-var-info=yes flags
gleipnir.out dropped from 557.6Mb to 93.1Mb or ~83.2% reduced file!)

The bigger the function with arguments it is called the bigger the
drop.

This also boosts mapvisual, as it does not have to store the function
while parsing it, as this information is no longer needed.